pyproject.toml
[tool.poetry]
name = "slypro-dz30"
version = "0.1.0"
description = ""
authors = ["flood <teranys1986@gmail.com>"]
readme = "README.md"
packages = [{include = "slypro_dz30"}]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
pillow = "^9.5.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
